making collection search a public widget
